add mappings
