+++ src/index.ts
@@ -1,6 +1,8 @@
 
 export { default as default } from './event'
-export { default as World, Block } from './world'
+export { default as World } from './world'
+export { default as Block } from "./block"
 export { default as Player } from './player'
 export * as Type from './types'
 export { MessageType } from './consts'
+export { BlockMappings, BlockMappingsReverse } from './mappings'
