+++ src/world.ts
@@ -1,4 +1,7 @@
 
+import Block from "./block"
+import { BlockMappings, BlockMappingsReverse } from './mappings'
+
 function get2dArray(width, height) {
     const arr = new Array(width)
     for (let i = 0; i < width; i++) {
@@ -8,9 +11,6 @@ function get2dArray(width, height) {
 }
 
 export default class World {
-    static mappings: {[keys: string]: number}
-    static reverseMapping: {[keys: number]: string}
-
     public width: number
     public height: number
     public foreground: Block[][]
@@ -23,21 +23,13 @@ export default class World {
         this.background = get2dArray(width, height)
     }
 
-    /**
-     * Set the block string to id mappings
-     */
-    public setMappings(map: {[keys: string]: number}) {
-        World.mappings = World.mappings || map
-        World.reverseMapping = World.reverseMapping || Object.fromEntries(Object.entries(map).map(a => a.reverse()))
-    }
-
     clear(border: boolean) {
         for (let x = 0; x < this.width; x++) {
             for (let y = 0; y < this.height; y++) {
                 const atBorder = border && (x == 0 || y == 0 || x == this.width - 1 || y == this.height - 1)
 
                 this.foreground[x][y] = atBorder ? new Block(World.mappings['basic_gray']) : new Block(0),
-                this.background[x][y] = new Block(0)
+                    this.background[x][y] = new Block(0)
             }
         }
     }
@@ -68,10 +60,10 @@ export default class World {
     deserializeBlock(buffer: Buffer, offset: number): [Block, number] {
         const id = buffer.readInt32LE(offset)
         const block = new Block(id)
-        
+
         offset += 4
 
-        switch (World.reverseMapping[id]) {
+        switch (BlockMappingsReverse[id]) {
             case 'empty':
                 return [new Block(0), offset]
 
@@ -84,7 +76,7 @@ export default class World {
 
             case 'portal':
                 block.rotation = buffer.readInt32LE(offset)
-                block.portal_id = buffer.readInt32LE(offset + 4) 
+                block.portal_id = buffer.readInt32LE(offset + 4)
                 block.target_id = buffer.readInt32LE(offset + 8)
                 return [block, offset + 12]
 
@@ -103,7 +95,7 @@ export default class World {
         const layer = l == 1 ? this.foreground : this.background
         const block = layer[x][y] = new Block(id)
 
-        switch (World.reverseMapping[id]) {
+        switch (BlockMappingsReverse[id]) {
             case 'coin_gate':
             case 'blue_coin_gate':
             case 'coin_door':
@@ -129,8 +121,8 @@ export default class World {
     }
 
     copy(x1: number, y1: number, x2: number, y2: number): World {
-        if (x2 < x1) {let tmp = x2; x2 = x1; x1 = tmp }
-        if (y2 < y1) {let tmp = y2; y2 = y1; y1 = tmp }
+        if (x2 < x1) { let tmp = x2; x2 = x1; x1 = tmp }
+        if (y2 < y1) { let tmp = y2; y2 = y1; y1 = tmp }
 
         const world = new World(x2 - x1 + 1, y2 - y1 + 1)
 
@@ -141,20 +133,5 @@ export default class World {
             }
 
         return world
-    }
-}
-
-export class Block {
-    public id: number
-    
-    constructor(id) {
-        if (typeof id == 'string') {
-            id = World.mappings[id]
-        }
-        this.id = id
-    }
-
-    get name() {
-        return World.reverseMapping[this.id]
     }
 }
